Main refacotring of startNewQuery computation: let DSCluster start it, without passing through DSInterface.
We now have one more ActorSystem with a ClusterManagerActor: its duties are:
* receive CreateChild for the first time (from DSCluster).
* forward in SendAll to ROBOT topic.
* receive MissionAccomplished messages on CLUSTERMANAGER topic,
  getting the MATCH / FAIL / DONTKNOW result.
* receive DeadLetters message. This may be useful to monitor the cluster.

The topic on which the QueryChecker listen is got from query.version,
initialized by DSCluster for the first time.

To know when the computation is over still in the DONTKNOW status,
we have a `left' parameter in TryNewQuery which is decremented at each forward,
initialized with the number of robots.